Added Component::VClone to the Component interface.
VClone creates a copy of the invoking Component. This is used by
GameObject to copy Components for which it doesn't have knowledge of the
underlying type. Added TestComponent::VClone and MeshRenderer::VClone to
ensure compatibility with the Component interface.

Removed ComponentFactory and ICloneable which were previous attempts to
solve this problem.