Fix for thirddegree/hatchitgame#29.
The issue that caused this was that Hatchit::Game::Component was a viable
template instantiation for Hatchit::Game::Component::GetComponentId<T>.
This caused problems in the JSON parsing code in Hatchit::Game::Scene
which deals primarily with pointers to Hatchit::Game::Components, all
Hatchit::Game::Components were being added with the same
Hatchit::Core::Guid because of this. Added a virtual VGetComponentId
function to the Hatchit::Game::Component class so that we can determine
the Core::Guid of an unknown Component. Added template specializations for
Hatchit::Game::GameObject::AddComponent and
Hatchit::Game::GameObject::AddUnitializedComponent, which accept a pointer
to a Hatchit::Game::Component and use VGetComponentId to appropriately
store the Hatchit::Game::Component. Updated all other templated
Hatchit::Game::GameObject functions to no longer accept
Hatchit::Game::Component as a valid template instantiation.